feat: GuardDuty Normalizer (#52)
This PR also renames the vendor names. We have been using SecurityHub as a vendor which is not appropriate because for our purpose, SecurityHub is more of a collection of vendors. For example, it can ingest alerts from Inspector, GuardDury, Config etc. As such, in this PR, we are breaking up SecurityHub into individual vendors (aws-config, guard duty etc)